app/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kali-tweaks==2023.3.2
MarkupSafe==2.1.3
numpy==1.26.2
python-apt==2.6.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
